Designed Enhanced ER diagram in lucid app
Every Dealership has at least one or many employees.
One employee allowed to work in one and only one Dealership. 
A vehicle has only one owner (supplier). 
An employee has none or many dependent.
There is only one and one Dealership in a city ( location )
The Dealership has at least one order. 
A customer has ordered at least one vehicle.
Employee have one location. ( they are not roomate as well ) 
Other suppliers include auctions, bank foreclosures, and private sellers except other [dealerships] (Cambridge, Kitchener, and London).
One Dealerships can be a supplier of another Dealership. For example, there is a Honda-Civic in the Kitchener branch, which a customer can order in the London branch. So Dealership in Kitchener can be the supplier of Honda civic for the London branch.  
There is one comprehensive list of vehicles of all branches, So other Dealerships can refer to it to finalize a customer's order. (inventory)
We have two relations, 
from Order -> Employee ->  Dealership, 
from Order -> Dealership. (Order table is kind of join table for customer and Dealership)
The two relations are that the Dealership should assign an employee to the order(s). In other words, when customers submit an order, it is not specified Which employee must be responsible for the order.
I consider inventory as a simulated vehicles warehouse, where all vehicles with all kinds of attributes ranging from price, location, model to the supplier discovered there. So employees can sell cars from other wally's Dealerships. For example, they can sell a vehicle located in London to a customer in Kitchener. Dealerships can refer to One comprehensive list to serve a customer. ( one dealership can be a supplier of another dealership )
Customer's order has status 1-completed, 2-failed, 3- in-progress. For example, there might be several orders for one vehicle (in progress) or several vehicles in one order (failed). That is why it is many to many (order->vehicle) relationship. A new record in the order table points out a specific vehicle and its status. Once the order's status is switched to completed, the system will remove it from inventory. 
Salespersons and managers are considered as an employee. Salespeople are employees, and managers are employees ( it has a relation to itself ).